OPEN - issue SEAMSOCIAL-12: Allow the support of multi-account on a given social service https://issues.jboss.org/browse/SEAMSOCIAL-12
Put all the services bean to dependent scope and created a qualifier
to allow bean selection for a given service. Thus it's now possible to
get more than one bean for each services.
The OAuthSettings is now in @ApplicationScope since we support only
one OAuth application by service. The config file will add qualifiers
to create more bean.

New Enum class with the list of supported service and the @RelatedTo
qualifier to bind a bean to a given service